VisualizerDeconstruct is the new Visualizer
